finished preliminary newMessage

+++ java/src/Messenger.java
@@ -17,6 +17,8 @@ import java.sql.Statement;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
+import java.sql.Timestamp;
+import java.util.Date;
 import java.io.File;
 import java.io.FileReader;
 import java.io.BufferedReader;
@@ -469,6 +471,7 @@ public List<List<String>> executeQueryResult (String query) throws SQLException 
                             System.out.print("\t");
                             printDashes(subTitle3.length());
                             System.out.println("\n");
+							NewMessage(esql, au);
 
                             break;
                         case 4: //settings
@@ -820,13 +823,161 @@ public List<List<String>> executeQueryResult (String query) throws SQLException 
    }//end
 */
 
-   public static void NewMessage(Messenger esql){
+   public static void NewMessage(Messenger esql, aUser au){
         try{
+			System.out.println("Enter The names of whome you wish to message(enter empty when done)");
+			ArrayList<String> reciv = new ArrayList<String>();
+			reciv.add(au.login);
+			int rtotal = 0;
+			String r = in.readLine();
+			while(!r.equals(""))
+			{
+				//check if usr exists
+				String ue = String.format("select * from usr where login = '%s'", r);
+				int uc = esql.executeQuery(ue);
+				if(uc == 0)
+				{
+					System.out.println("User does not exist");
+				}
+				else
+				{
+					// check if blocked
+					String cb = String.format("select * from USER_LIST_CONTAINS where list_member ='%s' and list_id = '%s' ", r, au.block_list);
+					int cc = esql.executeQuery(cb);
+					if(cc == 1)
+					{
+						System.out.println("He is blocked");
+					}
+					else
+					{
+						reciv.add(r);
+						rtotal = rtotal +1;
+					}
+					
+				}
+				r = in.readLine();
+			}
+			if(rtotal == 0)
+			{
+				System.out.println("Nobody to message back to menu");
+				return;	
+			}
+
+			//query for an existing chat
+			String ec = String.format("select chat_id from (select * from CHAT_LIST where member = '%s') as c group by c.chat_id having count(*) = '%s' ",au.login, rtotal );
+			List<List<String>> possible = esql.executeQueryResult(ec);
+			// check if any of the possbile have all the pl of the message as recipients
+			String hit = "";
+			for(int i = 0; i < possible.size() && hit.equals(""); i++)
+			{
+				int count = 1;
+				for(int j = 0; j <reciv.size() && count == 1; j++)
+				{
+					//query to see they exist in chat_list
+					String ecl = String.format("select * from CHAT_LIST where member = '%s' and chat_id = '%s'" , reciv.get(j), possible.get(i).get(0));
+					int ecn = esql.executeQuery(ecl);
+					if(ecn == 0)
+					{
+						count = 0;
+					}
+				}
+				if(count == 1)
+				{
+					hit = possible.get(i).get(0);
+				}
+			}
+			// 
+			if(hit.equals(""))
+			{
+				String type = "private";
+				if(reciv.size() > 2)
+				{
+					type = "group";
+				}
+				// didn't find chat make new chat
+				System.out.println("Making new chat ");
+				String eu = String.format("insert into CHAT(chat_type, init_sender) Values('%s', '%s')", type, au.login); 
+        		esql.executeUpdate(eu);
+         		int chat_id = esql.getCurrSeqVal("chat_chat_id_seq");
+				System.out.println("chat made");
+
+	
+				// add all
+				for(int x = 0; x < reciv.size(); x++)
+				{
+					String adda = String.format("insert into CHAT_LIST(chat_id,member) values('%s', '%s')", chat_id, reciv.get(x));
+					esql.executeUpdate(adda);
+					System.out.println("new chat_list made");
+				}
+				hit = String.format("'%s'",chat_id);
+			}
+			// here on hit is chat_id
+			
+				//get message
+			System.out.println("Type your message");
+			String msg = in.readLine();
+			while(msg.equals(""))
+			{
+				System.out.println("must have a msg");
+				msg = in.readLine();
+			}
+
+			//assume have msg now
+			// get if they want destr_timestamp
+			System.out.println("Do you want destruction time? (y/n)");
+			String ans2 = in.readLine();
+			while(!ans2.equals("y") && !ans2.equals("n"))
+			{
+				System.out.println("This is (y/n) only");
+				ans2 = in.readLine();
+			}
+			Timestamp tsd = null;
+
+			if(ans2.equals('y'))
+			{
+				System.out.println("How many milliseconds? (1hr:3,600,000) (min:60,000) (seconds:1000) :");
+				long mil = readChoice();
+				if( mil < 100)
+				{
+					System.out.println("minimum is 100 milliseconds");
+					mil = readChoice();
+				}
+				tsd = new Timestamp(mil);
+			}
+			//got destr if needed
+			Date date = new Date();
+			Timestamp ts = new Timestamp(date.getTime());
+
+			//make message
+			System.out.println("Making Message ");
+			if(tsd == null)
+			{
+				// TODO Temp fix need to come back 
+				tsd = new Timestamp((long) 111111111); 
+			}
+			String mm = String.format("insert into Message( msg_text, msg_timestamp, destr_timestamp,sender_login,chat_id) Values( '%s', '%s', '%s', '%s', '%s')", msg, ts, tsd, au.login, hit); 
+       		esql.executeUpdate(mm);
+
+			System.out.println("Message is made");
+			return;
+
+			
+			
 	    }catch (Exception e)
     	{
 	    	System.err.println(e.getMessage());
     	}
    }//end NewMessage
+
+// use to make more function clones
+ public static void temp(Messenger esql, aUser au){
+        try{
+			
+	    }catch (Exception e)
+    	{
+	    	System.err.println(e.getMessage());
+    	}
+   }//end temp
 
 
 public static void DeleteMessage(Messenger esql, aUser au, List<String> message)
@@ -1063,7 +1214,7 @@ public static void DeleteMessage(Messenger esql, aUser au, List<String> message)
 		List<List<String>> temp = null;
         try{
 			// Get all the chats user has membership of and find most current 
-				String c_time = String.format("select * from (select m.chat_id, Max(m.msg_timestamp) as msg_timestamp from MESSAGE m,(select chat_id from CHAT_LIST where member = '%s') as c where c.chat_id = m.chat_id group by m.chat_id ) as id_t order by id_t.msg_timestamp DESC", au.login);  
+				String c_time = String.format("select * from (select m.chat_id as chat_id, Max(m.msg_timestamp) as msg_timestamp from MESSAGE m,(select chat_id from CHAT_LIST, USER_LIST_CONTAINS ulc where member = '%s' and not member = ulc.list_member) as c where c.chat_id = m.chat_id group by m.chat_id ) as id_t order by id_t.msg_timestamp DESC", au.login);  
  
 	    	temp = esql.executeQueryResult(c_time);
 			System.out.println("");
@@ -1107,14 +1258,20 @@ public static void DeleteMessage(Messenger esql, aUser au, List<String> message)
 			//currently prints out all notifications in one go 
 			for(int i = 0; i < n_message_id.size(); i++)
 			{
-				String get_msg_i = String.format("select msg_text from MESSAGE where msg_id = '%s'", n_message_id.get(i).get(0));
+				String get_msg_i = String.format("select msg_text, sender_login from MESSAGE where msg_id = '%s'", n_message_id.get(i).get(0));
 				List< List<String>> msg_i = esql.executeQueryResult(get_msg_i);
-
+				
+				// check if blocked 
+				String cb = String.format("select * from USER_LIST_CONTAINS where list_id = '%s' and list_member ='%s'", au.block_list, msg_i.get(0).get(1));
+				int cn = esql.executeQuery(cb);
+				if(cn == 0)
+				{
 				System.out.println(msg_i.get(0).get(0));
 				//assuming it is now considered read delete from Notifications
 				String msg_d = String.format("delete from NOTIFICATION where user_login = '%s' and msg_id = '%s'", au.login, n_message_id.get(i).get(0));
 				System.out.println("\n " + msg_d);
 				//esql.executeUpdate(msg_d);
+				}
 			}
